+++ db/migrate/20150326093122_add_taggings_counter_cache_to_tags.rb
@@ -1,6 +1,9 @@
 class AddTaggingsCounterCacheToTags < ActiveRecord::Migration
   def self.up
-    add_column :tags, :taggings_count, :integer, default: 0
+    unless column_exists?(:tags, :taggings_count)
+      add_column :tags, :taggings_count, :integer, default: 0
+      @column_count_on_tags_added_by_kb = true
+    end
     ActsAsTaggableOn::Tag.reset_column_information
     ActsAsTaggableOn::Tag.find_each do |tag|
       ActsAsTaggableOn::Tag.reset_counters(tag.id, :taggings)
@@ -8,6 +11,11 @@ class AddTaggingsCounterCacheToTags < ActiveRecord::Migration
   end
 
   def self.down
-    remove_column :tags, :taggings_count
+    @column_count_on_tags_added_by_kb ||= false
+    
+    if @column_count_on_tags_added_by_kb and column_exists?(:tags, :taggings_count)
+      remove_column :tags, :taggings_count
+      @column_count_on_tags_added_by_kb = false
+    end
   end
 end
